500 - 600 Verified column delete after FK delete. (1.00)

+++ devtl-app/app/Http/Controllers/SchemaTableColumnController.php
@@ -8,6 +8,21 @@ class SchemaTableColumnController extends Controller
 {
     public function delete($schemaTableColumn)
     {
+        if (
+            $schemaTableColumn->foreignRelationships->count()
+            || $schemaTableColumn->primaryRelationships->count()
+        ) {
+            $alert = [
+                'class' => 'danger',
+                'message' => __('form.fk_relationship_exists_for_delete_column'),
+            ];
+
+            return response()->json([
+                    'status' => false,
+                    'html' => view('layouts.alert', compact('alert'))->render(),
+                ], 422);
+        }
+
         $schemaTableColumn->delete();
         return response()->json(['status' => true]);
     }
